requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
aiohttp
azure-functions
azure-identity
azure-mgmt-core
azure-mgmt-resourcegraph
azure-mgmt-subscription
msrestazure
azure-data-tables
azure-mgmt-resource
azure-storage-blob
azure.mgmt.support
simple-colors
